Рефакторинг и доработка кода для бд

+++ Example and tests/test_telebot.py
@@ -3,7 +3,7 @@ from telebot import types
 import os
 from dotenv import load_dotenv
 
-import telebot_db
+from db_scripts import bingo_db
 
 load_dotenv()
 token = os.getenv("BOT_TOKEN")
@@ -16,30 +16,22 @@ ids = [int(i) for i in raw_ids.split(",")] if raw_ids else []
 
 
 #Заглушка для бд
+maxamount = 2
 counters={}
 for i in range(1, 13):
     counters[f'item{i}']=[0,[]]
 
-@bot.message_handler(commands=['start'])
-def start(message):
-
-    #print(f"Пользователь {message.from_user.id} пишет сообщения")
-    bot.send_message(message.chat.id, "Участвуйте в бинго!", reply_markup=user_keyboard(message.from_user.id))
-
-@bot.message_handler(func=lambda message: message.text in ["Получить бинго","Принять участие в бинго","Войти в админку"])
-def handle_button(message):
-    if message.text == "Получить бинго":
-        welcome(message)
-    if message.text == "Принять участие в бинго":
-        print(f"Пользователь {message.from_user.id} принял участие в бинго")
-    if message.text == "Войти в админку" and message.from_user.id in ids:
-        bot.send_message(message.chat.id, "Да здравствует админка:", reply_markup=admin_keyboard())
-#for admin
-def admin_keyboard():
+def setting_bingo_keyboard():
     markup = types.ReplyKeyboardMarkup(resize_keyboard=True)
     markup.add(types.KeyboardButton("Добавить пункт"))
     markup.add(types.KeyboardButton("Узнать все пункты"))
     markup.add(types.KeyboardButton("Удалить пункт"))
+    markup.add(types.KeyboardButton("Сгенерировать новое бинго"))
+    markup.add(types.KeyboardButton("Выйти изи настройки бинго"))
+    return markup
+def admin_keyboard():
+    markup = types.ReplyKeyboardMarkup(resize_keyboard=True)
+    markup.add(types.KeyboardButton("Настроить бинго"))
     markup.add(types.KeyboardButton("Выйти из админки"))
     return markup
 def user_keyboard(user_id):
@@ -51,21 +43,41 @@ def user_keyboard(user_id):
         btn_admin_button = types.KeyboardButton('Войти в админку')
         markup.add(btn_admin_button)
     return markup
+@bot.message_handler(commands=['start'])
+def start(message):
+    #print(f"Пользователь {message.from_user.id} пишет сообщения")
+    bot.send_message(message.chat.id, "Участвуйте в бинго!", reply_markup=user_keyboard(message.from_user.id))
+
+
 #for admin
-@bot.message_handler(func=lambda message: message.text in ["Добавить пункт","Узнать все пункты", "Удалить пункт","Выйти из админки"])
+@bot.message_handler(func=lambda message: True)
 def handle_button(message):
-    if message.text == "Добавить пункт":
-        msg = bot.send_message(message.chat.id, "Введите название")
-        bot.register_next_step_handler(msg, add_item)
-    if message.text == "Удалить пункт":
-        msg = bot.send_message(message.chat.id, "Введите номер пункта")
-        bot.register_next_step_handler(msg, delete_item)
-    if message.text == "Узнать все пункты":
-        items = telebot_db.get_all_items()
-        text_items = "\n".join([f"{i}: {text}" for i, text in items])
-        bot.send_message(message.chat.id, text_items)
-    if message.text == "Выйти из админки":
-        bot.send_message(message.chat.id, "Возврат к бинго", reply_markup=user_keyboard(message.from_user.id))
+    if message.text == "Получить бинго":
+        welcome(message)
+    if message.text == "Принять участие в бинго":
+        print(f"Пользователь {message.from_user.id} принял участие в бинго")
+
+    if message.from_user.id in ids:
+        if message.text == "Войти в админку":
+            bot.send_message(message.chat.id, "Да здравствует админка:", reply_markup=admin_keyboard())
+        if message.text == "Добавить пункт":
+            msg = bot.send_message(message.chat.id, "Введите название")
+            bot.register_next_step_handler(msg, add_item)
+        if message.text == "Удалить пункт":
+            msg = bot.send_message(message.chat.id, "Введите номер пункта")
+            bot.register_next_step_handler(msg, delete_item)
+        if message.text == "Узнать все пункты":
+            items = telebot_db.get_all_items()
+            text_items = "\n".join([f"{i}: {text}" for i, text in items])
+            bot.send_message(message.chat.id, text_items)
+
+        if message.text == "Настроить бинго":
+            bot.send_message(message.chat.id, "Идём в настройку", reply_markup=setting_bingo_keyboard())
+        if message.text == "Выйти изи настройки бинго":
+            bot.send_message(message.chat.id, "Возвращаемся в админку", reply_markup=admin_keyboard())
+
+        if message.text == "Выйти из админки":
+            bot.send_message(message.chat.id, "Возврат к бинго", reply_markup=user_keyboard(message.from_user.id))
 
 #for admin
 def add_item(message):
@@ -80,10 +92,9 @@ def delete_item(message):
         bot.send_message(message.chat.id, 'Удалено!')
     except ValueError:
         bot.send_message(message.chat.id, 'Ошибка!')
-#bingo_button
+#bingo_btn
 def inline_keyboard():
     markup = types.InlineKeyboardMarkup()
-    maxamount = 3
     text_complete = "✅"
     buttons = []
     for i in range(1, 13):
@@ -99,49 +110,45 @@ def inline_keyboard():
 
     return markup
 #bingo
-@bot.message_handler(commands=['btn_msg'])
 def welcome(message):
     with open('bingo.jpg', 'rb') as photo_file:
         bot.send_photo(message.chat.id, photo_file, caption='TestCaption',reply_markup=inline_keyboard())
-#bingo
-@bot.callback_query_handler(func=lambda call: call.data[0:4] == 'item')
-def callback_inline(call):
-    item_id = call.data
-    user_id = call.from_user.id
-    maxamount = 3
-    if (user_id not in counters[item_id][1]) and(counters[item_id][0]!=maxamount):
-        counters[item_id][1].append(user_id)
-        counters[item_id][0]+=1
 
-    new_markup = inline_keyboard()
-    try:
-        bot.edit_message_reply_markup(
-            chat_id=call.message.chat.id,
-            message_id=call.message.message_id,
-            reply_markup=new_markup
-        )
-    except telebot.apihelper.ApiTelegramException as e:
-        # Если прилетело "сообщение не изменено", просто игнорим
-        if "message is not modified" not in e.description:
-            raise
-    bot.answer_callback_query(call.id)
-#bingo
-@bot.callback_query_handler(func=lambda call: call.data == 'update')
+#bingo_func
+@bot.callback_query_handler(func=lambda call: True)
 def callback_inline_update(call):
-    item_id = call.data
-    user_id = call.from_user.id
+    command = call.data
+    if command == 'update':
+        new_markup = inline_keyboard()
+        try:
+            bot.edit_message_reply_markup(
+                chat_id=call.message.chat.id,
+                message_id=call.message.message_id,
+                reply_markup=new_markup
+            )
+        except telebot.apihelper.ApiTelegramException as e:
+            # Если прилетело "сообщение не изменено", просто игнорим
+            if "message is not modified" not in e.description:
+                raise
+        bot.answer_callback_query(call.id)
+    if command[0:4]=='item':
+        item_id = call.data
+        user_id = call.from_user.id
+        if (user_id not in counters[item_id][1]) and (counters[item_id][0] != maxamount):
+            counters[item_id][1].append(user_id)
+            counters[item_id][0] += 1
 
-    new_markup = inline_keyboard()
-    try:
-        bot.edit_message_reply_markup(
-            chat_id=call.message.chat.id,
-            message_id=call.message.message_id,
-            reply_markup=new_markup
-        )
-    except telebot.apihelper.ApiTelegramException as e:
-        # Если прилетело "сообщение не изменено", просто игнорим
-        if "message is not modified" not in e.description:
-            raise
-    bot.answer_callback_query(call.id)
+        new_markup = inline_keyboard()
+        try:
+            bot.edit_message_reply_markup(
+                chat_id=call.message.chat.id,
+                message_id=call.message.message_id,
+                reply_markup=new_markup
+            )
+        except telebot.apihelper.ApiTelegramException as e:
+            # Если прилетело "сообщение не изменено", просто игнорим
+            if "message is not modified" not in e.description:
+                raise
+        bot.answer_callback_query(call.id)
 
 bot.infinity_polling()
